perf(商城): bug修复

+++ src/main/java/ltd/newbee/mall/controller/mall/MallShopCatController.java
@@ -2,8 +2,8 @@ package ltd.newbee.mall.controller.mall;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.toolkit.CollectionUtils;
-import ltd.newbee.mall.controller.base.BaseController;
 import ltd.newbee.mall.constant.Constants;
+import ltd.newbee.mall.controller.base.BaseController;
 import ltd.newbee.mall.core.entity.ShopCat;
 import ltd.newbee.mall.core.entity.vo.MallUserVO;
 import ltd.newbee.mall.core.entity.vo.MyCouponVO;
@@ -40,7 +40,7 @@ public class MallShopCatController extends BaseController {
     @GetMapping("shopCart/getUserShopCartCount")
     public R getUserShopCartCount(HttpSession session) {
         Object attribute = session.getAttribute(Constants.MALL_USER_SESSION_KEY);
-        int count = 0;
+        long count = 0;
         if (attribute == null) {
             return R.success().add("count", count);
         }
